refactor: TagNameFormを抽出して新規作成フォームを共通化
PlayerTagManager・SessionTagManagerの新規作成フォームを共通の
TagNameFormコンポーネントに統一。PlayerTagManagerはChildrenとして
TagColorPickerを渡す形で色選択を追加する。
